Working game
Core gameplay is functional. need to build out missions by city, unlocking cities based on mission completion, and use of the inventory system and using a rewards from missions to buy items to unlock missions based on equipment requirements as well. possibly consumable that get depleted as they are used for missions.